requirements.txt
streamlit==1.28.0
tensorflow-cpu==2.13.0
opencv-python-headless==4.8.1
scikit-learn==1.3.0
joblib==1.3.0
numpy==1.24.0
Pillow==10.0.0